% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genResults.R
\name{genResults}
\alias{genResults}
\title{Use scorenet network to predict scores}
\usage{
genResults(network, testData)
}
\value{
A data frame containing unscaled true game scores, predicted game scores and the error
}
\description{
First this selects the best rep of the nn object passed to the function.
It then uses this specific network on a set of test data to predict the
score in a game of cricket. The actual score should be in the 51st column of the test data,
the prior columns being the runrates for the corresponding over.
The function then returns the dataframe of results
}
