% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netAnalysis.R
\name{netAnalysis}
\alias{netAnalysis}
\title{Perform analysis on the neural network}
\usage{
netAnalysis(results)
}
\value{
A corrolation and a density plot
}
\description{
Uses corrolation to see how accurate the predictions of the neural network
were at predicting cricket scores. Produces a density plot of the errors
to see if they are normally distributed.
}
