% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreNet.R
\name{scoreNet}
\alias{scoreNet}
\title{Train a neural network}
\usage{
scoreNet(hiddenLayer, reps, alpha, trainData, formula)
}
\value{
Am object of class nn.
}
\description{
This function trains a neural network based on runrate data from a 50 over innings.
The parameters of the network are passed to the function which essentially acts as a wrapper
on the neuralnet package.
}
\examples{
scoreNet(c(10,5,3),1000,0.01,formula)
}
\seealso{
\code{\link[neuralnet:neuralnet]{neuralnet::neuralnet()}} which this function acts as a wrapper for
}
