% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Unscale a column of data}
\usage{
unscale(target, original)
}
\value{
A column of data that is no longer scaled
}
\description{
Used for turning unscaled scores into actual cricket scores. Target refers to the
object that is to be unscaled, and original is the data that scale() was applied to
in the first place
}
